+++ src/dependency/core/agrupation/entrypoint.py
@@ -1,5 +1,6 @@
 import logging
 import time
+from threading import Event
 from dependency.core.agrupation.plugin import Plugin
 from dependency.core.resolution.container import Container
 from dependency.core.resolution.resolver import InjectionResolver
@@ -15,10 +16,15 @@ class Entrypoint:
         injectables = [
             provider
             for plugin in plugins
-            for provider in plugin.resolve_providers(container)]
+            for provider in plugin.resolve_providers(container)
+        ]
 
         self.resolver: InjectionResolver = InjectionResolver(
             container=container,
-            injectables=injectables)
+            injectables=injectables
+        )
         self.resolver.resolve_dependencies()
         _logger.info(f"Application started in {time.time() - self.init_time} seconds")
+
+    def main_loop(self) -> None:
+        Event().wait()