+++ src/example/plugin/hardware/factory/products/productA.py
@@ -1,7 +1,8 @@
-from dependency_injector.wiring import Provide, inject
+from dependency_injector.wiring import inject
 from dependency.core import Product, product
 from example.plugin.hardware.factory.interfaces import Hardware
 from example.plugin.base.number import NumberService, NumberServiceComponent
+
 
 @product(
     imports=[
@@ -12,8 +13,7 @@ class HardwareA(Hardware, Product):
     @inject
     def doStuff(self,
             operation: str,
-            number: NumberService = Provide[NumberServiceComponent.reference],
+            number: NumberService = NumberServiceComponent.provider,
         ) -> None:
-        print(f"Injected NumberService into HardwareA: {NumberServiceComponent.reference}")
         random_number = number.getRandomNumber()
         print(f"HardwareA {random_number} works with operation: {operation}")