+++ src/example/plugin/hardware/factory/products/productB.py
@@ -1,4 +1,4 @@
-from dependency_injector.wiring import Provide, inject
+from dependency_injector.wiring import inject
 from dependency.core import Product, product
 from example.plugin.hardware.factory.interfaces import Hardware
 from example.plugin.base.string import StringService, StringServiceComponent
@@ -12,7 +12,7 @@ class HardwareB(Hardware, Product):
     @inject
     def doStuff(self,
             operation: str,
-            string: StringService = Provide[StringServiceComponent.reference],
+            string: StringService = StringServiceComponent.provider,
         ) -> None:
         random_string = string.getRandomString()
         print(f"HardwareB {random_string} works with operation: {operation}")