+++ tests/example/test_module.py
@@ -1,25 +1,42 @@
-from typing import Container
-from dependency.core import Container, Module, module
+from dependency.core import Container, InjectionResolver, Module, module
 from example.plugin.base.number.providers.fake import NumberService, NumberServiceComponent
+from example.plugin.base.string.providers.fake import StringService, StringServiceComponent
+from example.plugin.hardware.factory.providers.creatorA import HardwareFactory, HardwareFactoryComponent
+from example.plugin.hardware.observer.providers.publisherA import HardwareObserver, HardwareObserverComponent
+
 
 @module()
 class TestingModule(Module):
     pass
 
 def test_change_parent_and_resolve():
-    NumberServiceComponent.injection.change_parent(TestingModule.injection)
-    assert NumberServiceComponent.injection.parent == TestingModule.injection
-    assert NumberServiceComponent.injection in TestingModule.injection.childs
-    assert NumberServiceComponent.reference == "TestingModule.NumberServiceComponent"
+    for component in (
+        NumberServiceComponent,
+        StringServiceComponent,
+        HardwareFactoryComponent,
+        HardwareObserverComponent,
+    ):
+        component.injection.change_parent(TestingModule.injection)
+
+    assert HardwareFactoryComponent.injection.parent == TestingModule.injection
+    assert HardwareFactoryComponent.injection in TestingModule.injection.childs
+    assert HardwareFactoryComponent.injection.reference == "TestingModule.HardwareFactoryComponent"
 
     container = Container()
-    injectables = TestingModule.resolve_providers(container)
-    for injectable in injectables:
-        injectable.do_wiring(container)
-    assert NumberServiceComponent.injection.injectable in injectables
-    assert NumberServiceComponent.injection.injectable.is_resolved
+    resolver = InjectionResolver(
+        container=container,
+        injectables=TestingModule.resolve_providers(container),
+    )
+    resolver.resolve_dependencies()
+
+    assert HardwareFactoryComponent.injection.injectable in resolver.injectables
+    assert HardwareFactoryComponent.injection.injectable.is_resolved
+
+    hardware_factory: HardwareFactory = HardwareFactoryComponent.provide()
+    hardware_a = hardware_factory.createHardware("A")
+    hardware_a.doStuff("operation1")
 
     number_service1: NumberService = NumberServiceComponent.provide()
     number_service2: NumberService = NumberServiceComponent.provide()
-    assert number_service1.getRandomNumber() == 42
-    assert number_service2.getRandomNumber() == 43
+    assert number_service1.getRandomNumber() == 43
+    assert number_service2.getRandomNumber() == 44